cleanup + readme examples + begin HER refactor
