+++ baselines/her/__init__.py
@@ -1,0 +1,1 @@
+from baselines.her.her import HER
